requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-dotenv
neo4j
pydantic
sentence-transformers
numpy
openai
